+++ sql/core/src/main/scala/org/apache/spark/sql/execution/r/ArrowRRunner.scala
@@ -31,9 +31,10 @@ import org.apache.spark.api.r._
 import org.apache.spark.api.r.SpecialLengths
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.execution.ArrowColumnVector
 import org.apache.spark.sql.execution.arrow.{ArrowUtils, ArrowWriter}
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.vectorized.{ArrowColumnVector, ColumnarBatch, ColumnVector}
+import org.apache.spark.sql.vectorized.{ColumnarBatch, ColumnVector}
 import org.apache.spark.util.Utils
